[SPARK-XXXXX][SHUFFLE] Test: SEQ_NUM is a long to match setSeqNum parameter type
Addresses reviewer feedback (apache/spark#55620 review by @viirya):

`StreamingShuffleMessage.setSeqNum` takes a `long`, so every use of
SEQ_NUM in the test suite was implicitly widening an `int` to a `long`.
Changed the constant's declared type to `long` (`42L`) so it matches
the underlying field's type. All call sites work as-is — JUnit's
`assertEquals(long, long)` overload resolves cleanly and `setSeqNum`
accepts the `long` directly.

+++ common/network-common/src/test/java/org/apache/spark/network/shuffle/streaming/StreamingShuffleMessageSuite.java
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 public class StreamingShuffleMessageSuite {
 
-  private static final int SEQ_NUM = 42;
+  private static final long SEQ_NUM = 42L;
 
   // ---- helpers ---------------------------------------------------------------
